PECNet/requirements.txt
torch==1.4.0
numpy
pandas
matplotlib
ipython
pyyaml